requirements.txt
Django>=2.4
django-model-utils>=2.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
python-dotenv==0.12.0
# Additional requirements go here
celery>=4.4
psycopg2>=2.7,<3.0
dj-database-url==0.5.0

# for production
gunicorn==20.0.4
